requirements.txt
neo4j==5.15.0
openai==1.3.8
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
uuid==1.30